# specs/003-backend-implementation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Full-Stack Todo API
  description: Backend API for the Full-Stack Todo application
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      operationId: listTasks
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, all]
            default: all
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tag_ids
          in: query
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: deadline_from
          in: query
          schema:
            type: string
            format: date-time
        - name: deadline_to
          in: query
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          schema:
            type: string
          description: Search in title and description
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [created_at, deadline, priority, title]
            default: created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tasks]
      summary: Create task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /tasks/{task_id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/taskIdParam'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Tasks]
      summary: Update task
      operationId: updateTask
      parameters:
        - $ref: '#/components/parameters/taskIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      tags: [Tasks]
      summary: Delete task
      operationId: deleteTask
      parameters:
        - $ref: '#/components/parameters/taskIdParam'
      responses:
        '204':
          description: Task deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{task_id}/complete:
    post:
      tags: [Tasks]
      summary: Mark task as complete
      operationId: completeTask
      parameters:
        - $ref: '#/components/parameters/taskIdParam'
      responses:
        '200':
          description: Task marked complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Better Auth
  parameters:
    taskIdParam:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  schemas:
    Priority:
      type: string
      enum: [CRITICAL, HIGH, MEDIUM, LOW]
      default: MEDIUM

    # Task Schemas
    TaskBase:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
        description:
          type: string
          nullable: true
        priority:
          $ref: '#/components/schemas/Priority'
        deadline:
          type: string
          format: date-time
          nullable: true
        time_estimate:
          type: integer
          minimum: 0
          nullable: true
          description: Estimated minutes to complete
        project_id:
          type: string
          format: uuid
          nullable: true
    TaskCreate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          required: [title]
          properties:
            tag_ids:
              type: array
              items:
                type: string
                format: uuid
    TaskUpdate:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          properties:
            is_completed:
              type: boolean
    Task:
      allOf:
        - $ref: '#/components/schemas/TaskBase'
        - type: object
          required: [id, title, priority, is_completed, user_id, created_at, updated_at]
          properties:
            id:
              type: string
              format: uuid
            is_completed:
              type: boolean
            user_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            is_overdue:
              type: boolean
              readOnly: true
            days_until_deadline:
              type: integer
              nullable: true
              readOnly: true
            tags:
              type: array
              items:
                $ref: '#/components/schemas/TagSummary'
...
    TaskResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Task'
    TaskListResponse:
      type: object
      required: [success, data, pagination]
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Project Schemas
...
    TagSummary:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
          format: uuid
        label:
          type: string
        color:
          type: string
          nullable: true
...
    Pagination:
      type: object
      required: [offset, limit, total]
      properties:
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
              example: Task with ID xyz not found
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: BAD_REQUEST
              message: Invalid request body
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: RESOURCE_NOT_FOUND
              message: Resource not found
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: title is required
security:
  - bearerAuth: []